added option to clear launchservices (open with) and icon index

+++ 11_system_and_app_preferences/11e_defaults_open_with.sh
@@ -17,6 +17,93 @@
 
 ###
 
+# option for cleaning launchservices index
+echo ''
+read -p "do you want to clean the launchservices (open with) index and the icon cache after setting the new defaults for open with (y/N)? " CONT1
+CONT1="$(echo "$CONT1" | tr '[:upper:]' '[:lower:]')"    # tolower
+sleep 0.1
+echo ''
+
+if [[ "$CONT1" == "y" || "$CONT1" == "yes" ]]
+then
+
+    # function for reading secret string (POSIX compliant)
+    enter_password_secret()
+    {
+        # read -s is not POSIX compliant
+        #read -s -p "Password: " SUDOPASSWORD
+        #echo ''
+        
+        # this is POSIX compliant
+        # disabling echo, this will prevent showing output
+        stty -echo
+        # setting up trap to ensure echo is enabled before exiting if the script is terminated while echo is disabled
+        trap 'stty echo' EXIT
+        # asking for password
+        printf "Password: "
+        # reading secret
+        read -r "$@" SUDOPASSWORD
+        # reanabling echo
+        stty echo
+        trap - EXIT
+        # print a newline because the newline entered by the user after entering the passcode is not echoed. This ensures that the next line of output begins at a new line.
+        printf "\n"
+        # making sure builtin bash commands are used for using the SUDOPASSWORD, this will prevent showing it in ps output
+        # has to be part of the function or it wouldn`t be updated during the maximum three tries
+        #USE_PASSWORD='builtin echo '"$SUDOPASSWORD"''
+        USE_PASSWORD='builtin printf '"$SUDOPASSWORD\n"''
+    }
+    
+    # unset the password if the variable was already set
+    unset SUDOPASSWORD
+    
+    # making sure no variables are exported
+    set +a
+    
+    # asking for the SUDOPASSWORD upfront
+    # typing and reading SUDOPASSWORD from command line without displaying it and
+    # checking if entered password is the sudo password with a set maximum of tries
+    NUMBER_OF_TRIES=0
+    MAX_TRIES=3
+    while [ "$NUMBER_OF_TRIES" -le "$MAX_TRIES" ]
+    do
+        NUMBER_OF_TRIES=$((NUMBER_OF_TRIES+1))
+        #echo "$NUMBER_OF_TRIES"
+        if [ "$NUMBER_OF_TRIES" -le "$MAX_TRIES" ]
+        then
+            enter_password_secret
+            ${USE_PASSWORD} | sudo -k -S echo "" > /dev/null 2>&1
+            if [ $? -eq 0 ]
+            then 
+                break
+            else
+                echo "Sorry, try again."
+            fi
+        else
+            echo ""$MAX_TRIES" incorrect password attempts"
+            exit
+        fi
+    done
+    
+    # setting up trap to ensure the SUDOPASSWORD is unset if the script is terminated while it is set
+    trap 'unset SUDOPASSWORD' EXIT
+    
+    # replacing sudo command with a function, so all sudo commands of the script do not have to be changed
+    sudo()
+    {
+        ${USE_PASSWORD} | builtin command sudo -p '' -k -S "$@"
+        #${USE_PASSWORD} | builtin command -p sudo -p '' -k -S "$@"
+        #${USE_PASSWORD} | builtin exec sudo -p '' -k -S "$@"
+    }
+    
+    echo ''
+
+else
+    :
+fi
+
+###
+
 PLIST="$HOME/Library/Preferences/com.apple.LaunchServices/com.apple.launchservices.secure.plist"
 BUDDY=/usr/libexec/PlistBuddy
 
@@ -54,9 +141,6 @@ default_open_with=(
 #"LSHandlerContentType           public.comma-separated-values-text          org.libreoffice.script"        # .csv
 # openoffice
 #"LSHandlerContentType           public.comma-separated-values-text          org.openoffice.script"          # .csv
-
-
-echo ""
 
 for entry in "${default_open_with[@]}"
 do
@@ -137,10 +221,33 @@ do
     done
 done
 
-echo ''
-echo "cleaning launchservices (open with) index..."
-/System/Library/Frameworks/CoreServices.framework/Frameworks/LaunchServices.framework/Support/lsregister -kill -r -domain local -domain system -domain user 
-echo ''
+if [[ "$CONT1" == "y" || "$CONT1" == "yes" ]]
+then
+
+    #echo ''
+    echo "cleaning launchservices (open with) index..."
+    /System/Library/Frameworks/CoreServices.framework/Frameworks/LaunchServices.framework/Support/lsregister -kill -r -domain local -domain system -domain user
+    echo "rebuilding launchservices (open with) index..."
+    /System/Library/Frameworks/CoreServices.framework/Frameworks/LaunchServices.framework/Support/lsregister -seed -r -domain local -domain system -domain user 2>&1 | grep -v "registered plugin"
+    sleep 3
+    echo ''
+    
+    # it seems to be necessary to clean the icon cache after cleaning the launchservices (open with) index
+    # cleaning icon cache
+    echo "cleaning icon cache..."
+    sudo rm -rf /Library/Caches/com.apple.iconservices.store
+    sudo find /private/var/folders/ -name com.apple.dock.iconcache -exec rm -rf {} \; 2>/dev/null
+    sudo find /private/var/folders/ -name com.apple.iconservices -exec rm -rf {} \; 2>/dev/null
+    sleep 3
+    #sudo touch /Applications/*
+    #sleep 3
+    killall Dock
+    #killall Finder
+    echo ''
+
+else
+    :
+fi
 
 echo "done ;)"
 echo "the changes need a reboot to take effect..."
